+++ package.json
@@ -1,25 +1,27 @@
 {
 	"name": "portafolio",
 	"type": "module",
-	"version": "1.1.5",
+	"version": "1.1.8",
 	"scripts": {
 		"dev": "astro dev --host",
 		"start": "astro dev",
 		"build": "astro check && tsc --noEmit && astro build",
 		"preview": "astro preview",
-		"astro": "astro"
+		"astro": "astro",
+		"format": "biome format --write ./src ",
+		"lint": "biome check --apply ./src"
 	},
 	"dependencies": {
-		"@astrojs/check": "^0.2.1",
+		"@astrojs/check": "^0.3.1",
 		"@astrojs/react": "^3.0.4",
-		"@astrojs/sitemap": "^3.0.2",
+		"@astrojs/sitemap": "^3.0.3",
 		"@astrojs/tailwind": "^5.0.2",
 		"@emailjs/browser": "^3.11.0",
-		"@nextui-org/react": "^2.1.13",
-		"@types/react": "^18.2.33",
+		"@nextui-org/react": "^2.2.1",
+		"@types/react": "^18.2.34",
 		"@types/react-dom": "^18.2.14",
 		"@types/react-scroll": "^1.8.9",
-		"astro": "^3.4.0",
+		"astro": "^3.4.3",
 		"framer-motion": "^10.16.4",
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
@@ -33,7 +35,8 @@
 		"typescript": "^5.2.2"
 	},
 	"devDependencies": {
-		"@biomejs/biome": "^1.3.1",
+		"@biomejs/biome": "^1.3.3",
+		"prettier-plugin-tailwindcss": "^0.5.6",
 		"tailwindcss-animated": "^1.0.1"
 	}
 }
