+++ src/components/Certificates.astro
@@ -1,8 +1,8 @@
 ---
 import { Link } from '@nextui-org/react'
 import { Certificates as CertificateComponent } from '@utils/certificates'
-import { RiMapPinLine } from 'react-icons/ri'
-import { TbCertificate } from 'react-icons/tb'
+import { RiMapPinLine } from 'react-icons/ri/index'
+import { TbCertificate } from 'react-icons/tb/index'
 ---
 
 <section>
