+++ src/components/Contact/Label.tsx
@@ -1,0 +1,9 @@
+const Label = ({ name }: { name: string }) => {
+	return (
+		<label htmlFor={name} className="text-white font-nunito">
+			{name}
+		</label>
+	)
+}
+
+export default Label
