+++ src/components/Footer.astro
@@ -45,5 +45,5 @@ import {
       </a>
     </div>
   </div>
-  <span>V1.1.5</span>
+  <span>V1.1.8</span>
 </footer>
