+++ src/pages/index.astro
@@ -1,7 +1,7 @@
 ---
 import About from '@components/About.astro'
 import Certificates from '@components/Certificates.astro'
-import Contact from '@components/Contact'
+import Contact from '@components/Contact/Contact'
 import Hero from '@components/Hero.astro'
 import Knowledge from '@components/knowledge'
 import Project from '@components/Project'
